+++ api/src/main/java/com/github/skjolber/packing/api/StackConstraint.java
@@ -1,11 +1,13 @@
 package com.github.skjolber.packing.api;
 
+import java.math.BigDecimal;
+
 public interface StackConstraint {
 
 	boolean canAccept(Stackable stackable);
 
 	boolean accepts(Stack stack, Stackable stackable);
 
-	boolean supports(Stack stack, Stackable stackable, StackValue value, int x, int y, int z);
+	boolean supports(Stack stack, Stackable stackable, StackValue value, BigDecimal x, BigDecimal y, BigDecimal z);
 
 }
